rearrange the files
