+++ projects/summer-fi/constants/creation-blocks.js
@@ -2,6 +2,4 @@ const creationBlocks = {
   AutomationV1Contract: 14583413,
 };
 
-module.exports = {
-  creationBlocks,
-};
+module.exports = creationBlocks;
