+++ projects/summer-fi/constants/index.js
@@ -1,0 +1,11 @@
+const creationBlocks = require("./creation-blocks");
+const abi = require("./abi");
+const contracts = require("./contracts");
+const logsTopic = require("./logs-topic");
+
+module.exports = {
+  creationBlocks,
+  abi,
+  contracts,
+  logsTopic,
+};
