+++ projects/summer-fi/handlers/dpm-positions.js
@@ -1,5 +1,5 @@
 const BigNumber = require("bignumber.js");
-const { blockQuery } = require("../helper/http");
+const { blockQuery } = require("../../helper/http");
 
 const WETH_ADDRESS = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2";
