+++ projects/summer-fi/handlers/index.js
@@ -1,0 +1,9 @@
+const { automationTvl } = require("./automation-v1");
+const { dpmPositions } = require("./dpm-positions");
+const { makerTvl } = require("./maker-vaults");
+
+module.exports = {
+  automationTvl,
+  dpmPositions,
+  makerTvl,
+};
